requirements.txt
# Зависимости проекта
aiogram==3.17.0
asyncio==3.4.3
python-dotenv==1.0.1
logging==0.4.9.6
requests==2.32.3



